refactor: 장소 검색 도메인 로직 훅으로 분리
